chore: rename local_workspace tools to tool.py
